requirements.txt
anyio==3.6.2
autopep8==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.2
passlib==1.7.4
pycodestyle==2.10.0
pydantic==1.10.7
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.4
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0
